requirements.txt
langchain-openai>=0.0.5
python-dotenv>=1.0.0
rich>=13.7.0
typer>=0.9.0
markdown>=3.5.1
browser-use>=0.1.0 